converted ajax to typescript - 1st draft
